requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.1.4
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
django-widget-tweaks==1.4.8
elasticsearch==7.10.0
hiredis==1.1.0
hyperlink==21.0.0
idna==3.1
incremental==17.5.0
msgpack==1.0.2
paho-mqtt==1.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pytz==2020.4
screen==1.0.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.12.1
urllib3==1.26.2
xlwt==1.3.0
zope.interface==5.2.0
